+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### NEXT
 
+- Worker: Handle new STUN "NOMINATION`" attribute `0x0030` ([XXXX](https://github.com/versatica/mediasoup/pull/XXXX)).
+
 ### 3.20.9
 
 - Worker: Replace `uint64_t` hash with `TupleKey` in `TransportTuple` to avoid hash collisions ([1823](https://github.com/versatica/mediasoup/pull/1823)).