+++ rust/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Worker: Replace `uint64_t` hash with `TupleKey` in `TransportTuple` to avoid hash collisions ([1823](https://github.com/versatica/mediasoup/pull/1823)).
 - Worker: Fix `SeqManager::GetMaxOutput()` ([1840](https://github.com/versatica/mediasoup/pull/1840)).
+- Worker: Handle new STUN "NOMINATION`" attribute `0x0030` ([XXXX](https://github.com/versatica/mediasoup/pull/XXXX)).
 
 ### 0.22.8